refactor: move repositories to app/repositories and fix imports
Also fixed missing type imports and lint errors in common.ts
