+++ app/plugins/repositories.ts
@@ -1,8 +1,8 @@
 import type { Database } from '~/types/database.types'
-import { type DrinkLabelsRepository, createDrinkLabelsRepository } from '~/app/repositories/drinkLabelsRepository'
-import { type DrinkCountersRepository, createDrinkCountersRepository } from '~/app/repositories/drinkCountersRepository'
-import { type DrinksRepository, createDrinksRepository } from '~/app/repositories/drinksRepository'
-import { type UserSettingsRepository, createUserSettingsRepository } from '~/app/repositories/userSettingsRepository'
+import { type DrinkLabelsRepository, createDrinkLabelsRepository } from '~/repositories/drinkLabelsRepository'
+import { type DrinkCountersRepository, createDrinkCountersRepository } from '~/repositories/drinkCountersRepository'
+import { type DrinksRepository, createDrinksRepository } from '~/repositories/drinksRepository'
+import { type UserSettingsRepository, createUserSettingsRepository } from '~/repositories/userSettingsRepository'
 
 export default defineNuxtPlugin(() => {
   const supabase = useSupabaseClient<Database>()
