+++ store/pages/drinks/index/state.ts
@@ -1,3 +1,5 @@
+import type { DrinkRow as Drink } from '~/repositories/drinksRepository'
+
 const deleteTarget = ref<Drink | null>(null)
 const showDeleteModal = ref<boolean>(false)
