+++ store/pages/labels/index/state.ts
@@ -1,3 +1,5 @@
+import type { DrinkLabelWithDefaultColor as DrinkLabel } from '~/repositories/drinkLabelsRepository'
+
 // 削除対象のDrinkデータ
 const deleteTarget = ref<DrinkLabel | null>(null)
 // 削除モーダル表示フラグ
