+++ store/supabase/state.ts
@@ -1,4 +1,4 @@
-import type { Database } from '~/database.types'
+import type { Database } from '~/types/database.types'
 
 export function useSupabaseState () {
   const supabase = useSupabaseClient<Database>()
